refactor: MemberRepository 인터페이스면 변경 #8
